+++ com.vmturbo.analysis/platform.analysis/src/main/java/com/vmturbo/platform/analysis/economy/UnmodifiableEconomy.java
@@ -226,7 +226,7 @@ public interface UnmodifiableEconomy {
     /**
      * @return balance account map associates with the {@link Economy}
      */
-    Map<Integer, BalanceAccount> getBalanceAccountMap();
+    Map<Long, BalanceAccount> getBalanceAccountMap();
 
     /**
      * save the {@link Topology} associated with this {@link Economy}
